backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
cryptography==41.0.7

# Real-time Communication
Flask-SocketIO==5.3.6
python-socketio==5.11.0
eventlet==0.35.1

# API Integration
requests==2.31.0
spotipy==2.23.0
urllib3==2.1.0

# Background Tasks
celery==5.3.5
redis==5.0.1
flower==2.0.1

# File Storage
boto3==1.34.14
Pillow==10.2.0

# Data Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Rate Limiting
Flask-Limiter==3.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
pydantic==2.5.3
email-validator==2.1.0

# Geolocation
geopy==2.4.1
haversine==2.8.1

# Development
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
pre-commit==3.6.0

# Production
gunicorn==21.2.0
gevent==23.9.1
